Added generics to the cuda functions

+++ src/cuda_functions/cuda_malloc.rs
@@ -23,7 +23,7 @@ mod tests {
     fn test_cuda_malloc() {
         let mut src = std::ptr::null_mut(); // This is almost never what you want
                                             // https://stackoverflow.com/questions/47878236/how-do-i-make-the-equivalent-of-a-c-double-pointer-in-rust
-        let _ = cuda_malloc(&mut src, 10 * std::mem::size_of::<i32>());
+        let _ = cuda_malloc(&mut src, 10 * std::mem::size_of::<f64>());
 
         assert!(!src.is_null()); // assert src is not null means the memory allocation is successful
